Api_Check/requirements.txt
annotated-types==0.7.0
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
config==0.5.1
idna==3.10
iniconfg==2.0.0
packaging==24.2
pip==25.0.1
pluggy==1.5.0
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic-core==2.27.1
python-dotenv==1.0.1
requests==2.32.3
typing_extensions==4.12.2
urllib3==2.3.0